Fix config issue for typescript language server
VSCode typescript language server looks exclusively for `tsconfig.json`
This prevents using the `tsconfig.json` for build processes and
`tsconfig.eslint.json` for eslint to lint files that will not be include
in the compiled build. It specifically throws errors when it can't
resolve json files when they are used in the test directory because it
is not included in `tsconfig.json`.

I fixed this by renaming the files, but it feels like a hack:
    - `tsconfig.json` -> `tsconfig.build.json`
    - `tsconfig.eslint.json` -> `tsconfig.json`

I think this could all be resolved in `.eslintrc` by removing the
`parserOptions.project` field. This field is required by the
`@typescript-eslint/recommended-requiring-type-checking` plugin which
provided a few bonus lint rules since it requires type checking. This
removal would obviously give up any beneficial linting supplied by the
package. I need to investigate how beneficial these rules are.